Corrected color assignment to the led's
I encountered an error where the time zone on raspberry pi this is running on was 3 hours ahead. Therefore, at the end of the week at night, there would be no data to display. Added a reminder to check print statement to check this. Additionally, I did not assign the led's their color correctly, so the program would throw an error.